double floating point precision

+++ ft_calib.py
@@ -10,7 +10,7 @@ def load_calib_matrix(calib_file):
         xmldoc = minidom.parse(calib_file)
     except IOError:
         print("\nCalib File %s could not be found. Returning identity matrix...\n"%calib_file)
-        return np.eye(6)
+        return np.eye(6,dtype=np.float64)
 
     itemlist = xmldoc.getElementsByTagName('UserAxis')
 
@@ -18,9 +18,9 @@ def load_calib_matrix(calib_file):
     for s in itemlist:
 
         parsed_mat = (s.attributes['values'].value).split()
-        mat.append([float(i) for i in parsed_mat])
+        mat.append([np.float64(i) for i in parsed_mat])
 
-    calib_mat = np.asarray(mat)
+    calib_mat = np.asarray(mat,dtype=np.float64)
 
     return calib_mat
